add a separate class AuthenticationTokenValidator for validating tokens
