+++ python/ray/includes/rpc_token_authentication.pxi
@@ -3,6 +3,7 @@ from ray.includes.rpc_token_authentication cimport (
     GetAuthenticationMode,
     CAuthenticationToken,
     CAuthenticationTokenLoader,
+    CAuthenticationTokenValidator,
 )
 from ray._private.authentication.authentication_constants import AUTHORIZATION_HEADER_NAME
 import logging
@@ -38,13 +39,18 @@ def validate_authentication_token(provided_token: str) -> bool:
     Returns:
         bool: True if token is valid, False otherwise
     """
+    cdef optional[CAuthenticationToken] expected_opt = CAuthenticationTokenLoader.instance().GetToken()
+
+    if not expected_opt.has_value():
+        return False
+
     # Parse provided token from Bearer format
     cdef CAuthenticationToken provided = CAuthenticationToken.FromMetadata(provided_token.encode())
 
     if provided.empty():
         return False
 
-    return CAuthenticationTokenLoader.instance().ValidateToken(provided)
+    return CAuthenticationTokenValidator.instance().ValidateToken(expected_opt, provided)
 
 
 class AuthenticationTokenLoader:
