+++ src/ray/rpc/authentication/authentication_token_loader.h
@@ -18,21 +18,12 @@
 #include <mutex>
 #include <optional>
 #include <string>
-#include <unordered_map>
 
 #include "ray/rpc/authentication/authentication_mode.h"
 #include "ray/rpc/authentication/authentication_token.h"
-#include "ray/rpc/authentication/k8s_util.h"
 
 namespace ray {
 namespace rpc {
-
-// Hash function for AuthenticationToken
-struct AuthenticationTokenHash {
-  std::size_t operator()(const AuthenticationToken &token) const {
-    return std::hash<std::string>()(token.ToValue());
-  }
-};
 
 /// Singleton class for loading and caching authentication tokens.
 /// Supports loading tokens from multiple sources with precedence:
@@ -54,14 +45,6 @@ class AuthenticationTokenLoader {
   /// Caches the token if it loads it afresh.
   /// \return true if a token exists, false otherwise.
   bool HasToken();
-
-  /// Validate the provided authentication token.
-  /// For TOKEN mode, it compares with the loaded token.
-  /// For K8S mode, it uses Kubernetes TokenReview and SubjectAccessReview APIs.
-  /// The results for K8S mode are cached.
-  /// \param provided_token The token to validate.
-  /// \return true if the token is valid, false otherwise.
-  bool ValidateToken(const AuthenticationToken &provided_token);
 
   void ResetCache() {
     std::lock_guard<std::mutex> lock(token_mutex_);
@@ -89,15 +72,6 @@ class AuthenticationTokenLoader {
 
   std::mutex token_mutex_;
   std::optional<AuthenticationToken> cached_token_;
-
-  // Cache for K8s tokens.
-  struct K8sCacheEntry {
-    bool allowed;
-    std::chrono::steady_clock::time_point expiration;
-  };
-  std::mutex k8s_token_cache_mutex_;
-  std::unordered_map<AuthenticationToken, K8sCacheEntry, AuthenticationTokenHash>
-      k8s_token_cache_;
 };
 
 }  // namespace rpc
